Refactor i18n
