--- designer/client/i18n/index.js
@@ -1,1 +1,0 @@
-export { withI18n, i18n, I18n, initI18n, withI18nRef } from "./i18n";
